Move Admin Inventory to Pages folder
